@@ -13,10 +13,10 @@
     }
   ],
   "require": {
-    "php": ">=5.6.0"
+    "php": ">=7.2.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^4.8"
+    "phpunit/phpunit": "^8.0"
   },
   "suggest": {
   },
